add modal_context.tsx to resolve circular references
