+++ src/components/login_page_body/login_page_body.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import Image from 'next/image';
-import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import AvatarSVG from '@/components/avater_svg/avater_svg';
 import { ILoginPageProps } from '@/interfaces/page_props';
 import { Provider } from '@/constants/provider';
@@ -57,7 +57,7 @@ const Loader = React.memo(() => {
 
 const LoginPageBody = ({ invitation, action }: ILoginPageProps) => {
   const { t } = useTranslation('common');
-  const { toastHandler } = useGlobalCtx();
+  const { toastHandler } = useModalContext();
   const { isAuthLoading, authenticateUser, userAgreeResponse } = useUserCtx();
 
   useEffect(() => {
