+++ src/components/preview_invoice_modal/preview_invoice_modal.tsx
@@ -5,8 +5,7 @@ import { PiShareFat } from 'react-icons/pi';
 import { timestampToString } from '@/lib/utils/common';
 import { IPreviewInvoiceModal } from '@/interfaces/preview_invoice_modal';
 import { Button } from '@/components/button/button';
-// eslint-disable-next-line import/no-cycle
-import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import { MessageType } from '@/interfaces/message_modal';
 import { useTranslation } from 'next-i18next';
 import { ToastType } from '@/interfaces/toastify';
@@ -24,7 +23,8 @@ const PreviewInvoiceModal = ({
 }: IPreviewInvoiceModalProps) => {
   const { t } = useTranslation(['common', 'journal']);
   const { date, imgStr } = previewInvoiceModalData;
-  const { messageModalVisibilityHandler, messageModalDataHandler, toastHandler } = useGlobalCtx();
+  const { messageModalVisibilityHandler, messageModalDataHandler, toastHandler } =
+    useModalContext();
 
   // Info: (20240508 - Julian) 關閉 modal
   const closeClickHandler = () => modalVisibilityHandler();
