+++ src/components/project_roi_comparison_chart/project_roi_comparison_chart.tsx
@@ -18,6 +18,7 @@ import {
 import APIHandler from '@/lib/utils/api_handler';
 import { APIName } from '@/constants/api_connection';
 import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import { ToastType } from '@/interfaces/toastify';
 import { useUserCtx } from '@/contexts/user_context';
 import { LayoutAssertion } from '@/interfaces/layout_assertion';
@@ -167,7 +168,8 @@ const ProjectRoiComparisonChart = () => {
   const { t } = useTranslation(['common', 'project']);
   const { isAuthLoading, selectedCompany } = useUserCtx();
   const hasCompanyId = isAuthLoading === false && !!selectedCompany?.id;
-  const { toastHandler, layoutAssertion } = useGlobalCtx();
+  const { layoutAssertion } = useGlobalCtx();
+  const { toastHandler } = useModalContext();
 
   const minDate = new Date(DUMMY_START_DATE);
   const maxDate = new Date();
