+++ src/components/team_setting_modal/team_setting_modal.tsx
@@ -4,8 +4,7 @@ import APIHandler from '@/lib/utils/api_handler';
 import { useUserCtx } from '@/contexts/user_context';
 import { APIName } from '@/constants/api_connection';
 import { ICompany } from '@/interfaces/company';
-// eslint-disable-next-line import/no-cycle
-import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import { ToastType } from '@/interfaces/toastify';
 import { useTranslation } from 'next-i18next';
 
@@ -18,7 +17,7 @@ interface ITeamSettingModal {
 const TeamSettingModal = ({ isModalVisible, modalVisibilityHandler }: ITeamSettingModal) => {
   const { t } = useTranslation(['common', 'setting']);
   const { selectedCompany, selectCompany } = useUserCtx();
-  const { toastHandler } = useGlobalCtx();
+  const { toastHandler } = useModalContext();
   const [companyName, setCompanyName] = useState<string>(selectedCompany?.name ?? '');
 
   const {
