+++ src/pages/_app.tsx
@@ -9,6 +9,7 @@ import { GlobalProvider } from '@/contexts/global_context';
 import { DashboardProvider } from '@/contexts/dashboard_context';
 import { AccountingProvider } from '@/contexts/accounting_context';
 import { SessionProvider } from 'next-auth/react';
+import { ModalProvider } from '@/contexts/modal_context';
 import '@/styles/globals.css';
 
 function App({ Component, pageProps: { session, ...pageProps } }: AppProps) {
@@ -21,9 +22,11 @@ function App({ Component, pageProps: { session, ...pageProps } }: AppProps) {
           <UserProvider>
             <DashboardProvider>
               <AccountingProvider>
-                <GlobalProvider>
-                  <Component {...pageProps} />
-                </GlobalProvider>
+                <ModalProvider>
+                  <GlobalProvider>
+                    <Component {...pageProps} />
+                  </GlobalProvider>
+                </ModalProvider>
               </AccountingProvider>
             </DashboardProvider>
           </UserProvider>
