+++ src/pages/users/reports/analyses/view/[reportId].tsx
@@ -7,7 +7,7 @@ import Head from 'next/head';
 import React, { useEffect } from 'react';
 import APIHandler from '@/lib/utils/api_handler';
 import { APIName } from '@/constants/api_connection';
-import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import { ToastType } from '@/interfaces/toastify';
 import ViewAnalysisSection from '@/components/view_analysis_section/view_analysis_section';
 import { ReportLanguagesKey } from '@/interfaces/report_language';
@@ -39,7 +39,7 @@ const ViewAnalysisReportPage = ({
   endTimestamp,
 }: IServerSideProps) => {
   const { t } = useTranslation('common');
-  const { toastHandler } = useGlobalCtx();
+  const { toastHandler } = useModalContext();
   const { selectedCompany, isAuthLoading } = useUserCtx();
   const hasCompanyId = isAuthLoading === false && !!selectedCompany?.id;
   const [reportData, setReportData] = React.useState<IReportOld>({
