+++ src/pages/users/reports/financials/view/[reportId].tsx
@@ -12,7 +12,7 @@ import Head from 'next/head';
 import React, { useEffect } from 'react';
 import APIHandler from '@/lib/utils/api_handler';
 import { APIName } from '@/constants/api_connection';
-import { useGlobalCtx } from '@/contexts/global_context';
+import { useModalContext } from '@/contexts/modal_context';
 import { ToastType } from '@/interfaces/toastify';
 import { useUserCtx } from '@/contexts/user_context';
 import { DEFAULT_SKELETON_COUNT_FOR_PAGE } from '@/constants/display';
@@ -28,7 +28,7 @@ interface IServerSideProps {
 
 const ViewFinancialReportPage = ({ reportId, reportType }: IServerSideProps) => {
   const { t } = useTranslation('common');
-  const { toastHandler } = useGlobalCtx();
+  const { toastHandler } = useModalContext();
   const { selectedCompany, isAuthLoading } = useUserCtx();
   const hasCompanyId = isAuthLoading === false && !!selectedCompany?.id;
   const [reportData] = React.useState<IReportOld>({
